requirements.txt
Django==5.1.7
djangorestframework==3.15.2
requests==2.32.3 
djangorestframework-simplejwt==5.3.1
pillow==10.4.0 